change int to size_t

+++ include/interface/functions.hpp
@@ -9,7 +9,7 @@ bool hydrophones(Data*, const std::string, const std::string);
 
 bool vision(Data*, const std::string, const std::string, size_t, size_t);
 
-bool camera(Data*, const std::string, const std::string, int);
+bool camera(Data*, const std::string, const std::string, size_t);
 
 bool mission(Data*, const std::string, const std::string);
 bool modeling(Data*, const std::string, const std::string);
